+++ packages/manager/src/hooks/usePendingUpload.ts
@@ -1,0 +1,7 @@
+import { useSelector } from 'react-redux';
+
+import { ApplicationState } from 'src/store';
+
+export const usePendingUpload = () => {
+  return useSelector((state: ApplicationState) => state.pendingUpload);
+};
